Fix SH8601 AMOLED display initialization
- Add AXP2101 PMU driver to enable power rails for display
- Use Waveshare's modified GFX library with proper SH8601 support
- Fix I2C pins (GPIO15/14 instead of GPIO8/9)
- Add TCA9554 reset sequence and brightness control

The display now shows content correctly. The upstream Arduino_GFX
library has different SH8601 initialization than what this hardware
requires - Waveshare's modified version is needed.

+++ src/touch.cpp
@@ -54,7 +54,7 @@ static bool ft_read_reg(uint8_t reg, uint8_t *data, uint8_t len) {
         return false;
     }
 
-    Wire.requestFrom(TOUCH_I2C_ADDR, len);
+    Wire.requestFrom((uint8_t)TOUCH_I2C_ADDR, (uint8_t)len);
     for (uint8_t i = 0; i < len && Wire.available(); i++) {
         data[i] = Wire.read();
     }
